+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.6] - 2021-03-06
 ### Changed
 - packages/sdk
   - Update: Remove keys.updated event handler since that path was not supported yet and could derive on mem leaks
@@ -74,7 +76,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - packages/cli
   - Pretty print logs
 
-[Unreleased]: https://github.com/geut/permanent-seeder/compare/v1.0.5...HEAD
+[Unreleased]: https://github.com/geut/permanent-seeder/compare/v1.0.6...HEAD
+[1.0.6]: https://github.com/geut/permanent-seeder/compare/v1.0.5...v1.0.6
 [1.0.5]: https://github.com/geut/permanent-seeder/compare/v1.0.4...v1.0.5
 [1.0.4]: https://github.com/geut/permanent-seeder/compare/v1.0.3...v1.0.4
 [1.0.3]: https://github.com/geut/permanent-seeder/compare/v1.0.2...v1.0.3